+++ src/main.zig
@@ -18,6 +18,9 @@ comptime {
     _ = exports.turbotoken_count_non_ascii_utf8_neon;
     _ = exports.turbotoken_count_non_ascii_utf8_dotprod;
     _ = exports.turbotoken_count_non_ascii_utf8_sme;
+    _ = exports.turbotoken_count_ascii_class_boundaries_utf8;
+    _ = exports.turbotoken_count_ascii_class_boundaries_utf8_scalar;
+    _ = exports.turbotoken_count_ascii_class_boundaries_utf8_neon;
     _ = exports.turbotoken_encode_utf8_bytes;
     _ = exports.turbotoken_decode_utf8_bytes;
     _ = exports.turbotoken_encode_bpe_from_ranks;